Tried refactoring Post/Repost/Quotes into ArticleSkeleton
